Record a high-water-mark of "dirty" lines in the console screen buffer and avoid sending lines in the window that aren't dirty yet.
The idea is that the bottom of the console screen commonly has blank lines
that do not need to be sent to the client, because nothing has been written
there yet.  For example, when the console first appears, there is typically
a command-line prompt at the top of the window.  When a console is expanded
vertically, blank lines commonly appear at the bottom.

Other misc changes to the output:
 - When we send one later, also send all following lines.
 - When we save a line into m_bufferData, fill in the rest of the line with
   spaces.  This way, it's possible to expand the console horizontally
   without the agent resending all the lines.
 - If nothing on the console has changed, then leave the cursor alone.
   Don't hide, show, or move it.