Start the agent in a new window station.
The intent is to avoid a problem on Windows XP.  On that OS, sending
a "Select All" command to freeze one console window also interrupts any
"window activity" (e.g. moving/closing/scrolling/right-clicking/...) on
all other console windows.

My hope is that putting the console into a special window station will
isolate it enough to avoid this problem.  I noticed that the problem didn't
happen when I ran the TestNetServer as a service, which ran both the
TestNetServer and Agent processes with a special window station.

Note that even though the Agent runs in a special window station, the shell
and its children run in the normal window station.