Consolidate GetVersionEx calls and silence the MSVC deprecation warning
MSVC complained that GetVersionEx was deprecated when the /SDL checks were
enabled.

Comment added to WindowsVersion.cc:

Allow use of deprecated functions (i.e. GetVersionEx).  We need to use
GetVersionEx for the old MinGW toolchain and with MSVC when it targets XP.
Having two code paths makes code harder to test, and it's not obvious how
to detect the presence of a new enough SDK.  (Including ntverp.h and
examining VER_PRODUCTBUILD apparently works, but even then, MinGW-w64 and
MSVC seem to use different version numbers.)