Change the default way winpty enters the terminal's mouse mode.
By default, winpty now puts a new console into QuickEdit mode.  If the
agents detects that the console has disabled QuickEdit and enabled
ENABLE_MOUSE_INPUT, then it puts the terminal into mouse mode.

This convention doesn't work with all Windows programs -- it does work
with Far Manager and "winpty-agent.exe --show-input --with-mouse".  The
practice is described in one of Raymond Chen's posts[1].

For other programs, winpty.exe still has a --mouse option, and I wrote a
script in misc/ConinMode.ps1 that controls the QuickEdit (and InsertMode)
flags from the command-line, which is useful, because the console
properties dialog is inaccessible from within winpty.

[1] https://blogs.msdn.microsoft.com/oldnewthing/20130506-00/?p=4453