Improve debugging output in Win32ConsoleBuffer::read
The reads were failing when running wmic.exe under winpty on Windows
10 build 14342.  In that build, after running wmic.exe, the console
buffer was resized to (1500,300), but the window was not moved upward,
so it was hundreds to thousands of lines below the actual buffer.

This is obviously a Windows 10 bug, because even running wmic.exe in a
normal console demonstrates the issue.  Simply fill the console with more
than 300 lines of junk, then try to run wmic.exe, and none of the
wmic.exe output is visible, even after hitting Return repeatedly.